+++ src/components/Item/Item.js
@@ -1,17 +1,22 @@
 import React from 'react';
 
 export const Item = ({item})  => {
-    const item = [{
-        id: '1', 
-        title: "IRON MAN", 
-        duration: "3 hrs",
-        price: 8.000,
-        pictureUrl: "https://prod-ripcut-delivery.disney-plus.net/v1/variant/disney/5B5F22E0E37CF72BA1D2E1A8778BA6100C4ED9AC22245DFD8C84AE814DC63F8B/scale?width=400&aspectRatio=1.78&format=jpeg",
-        año: "2020" }]
+
+        const addZero = (num) => {
+            let number = num.toString();
+            let decimal = (number.length < 2) ? `${number}.000` :
+            (number.length < 3) ? `${number}.00` :
+            `${number}`;
+            return decimal;   
+        }
 
   return (
-    <>
-    <p>vista de producto al darle click </p>
-    </>
+        <div key={item.id}>
+            <a>
+                <img src={item.pictureUrl} />
+            </a>
+            <p className="name-card">{item.title } ({item.duration})</p>
+            <p>Price: $ {addZero(item.price) }</p>
+        </div>
   ); 
 }
